Experimental: use usize instead of bochs types

+++ src/config.rs
@@ -1,1 +1,2 @@
-pub type BxPhyAddress = u64;
+// pub type BxPhyAddress = u64;
+pub type BxPhyAddress = usize;
